Split wgt-size-offset crate impl
