+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Unreleased
 
+* Fix `Length::Default` having a layout effect in exact size properties. !!: TODO
+    - This default value now *disables* the properties, the required behavior for integration with `when` blocks in widgets.
+    - Changed properties: `layout::{size, max_size, width, height, max_width, max_height}`.
 * Fix buttons at the top-level of `Menu!` not returning focus on click.
 * View-process now tries to guess image format in case of header decode error for an extension or mime defined format.
+* Fix `cargo zng fmt` wrap instability inside macros.
 
 # 0.17.0
